+++ CHANGELOG.md
@@ -27,6 +27,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-3690 Update basal tooltip style in Daily view
 - YLP-3700 Update time change tooltip style in Daily view
 - YLP-3702 Update Time In Loop Mode stat style
+- YLP-3707 User account page new design
 - YLP-3710 Move the "EVENTS" section above the glycemia curve
 
 ### Engineering